requirements.txt
accelerate==1.2.1
aiofiles==23.2.1
aiohappyeyeballs==2.4.8
aiohttp==3.11.13
aiosignal==1.3.2
airportsdata==20241001
annotated-types==0.7.0
anyio==4.8.0
astor==0.8.1
asttokens==3.0.0
attrs==25.1.0
audioread==3.0.1
av==14.2.0
bitsandbytes==0.45.3
blake3==1.0.4
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
comm==0.2.2
compressed-tensors==0.8.1
contourpy==1.3.1
cut-cross-entropy==25.1.1
cycler==0.12.1
datasets==3.2.0
debugpy==1.8.11
decorator==5.2.1
depyf==0.18.0
diffusers==0.32.2
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
docstring_parser==0.16
einops==0.8.1
eval_type_backport==0.2.2
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.11
ffmpy==0.5.0
filelock==3.17.0
fire==0.7.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.9.0
gguf==0.10.0
gradio==5.12.0
gradio_client==1.5.4
h11==0.14.0
hf_transfer==0.1.9
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.29.1
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.0.0
interegular==0.3.3
ipykernel==6.29.5
ipython==8.31.0
jedi==0.19.2
jieba==0.42.1
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
lark==1.2.2
lazy_loader==0.4
librosa==0.10.2.post1
llamafactory==0.9.2.dev0
llvmlite==0.44.0
lm-format-enforcer==0.10.9
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistral_common==1.5.2
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.19.0
multidict==6.1.0
multiprocess==0.70.16
nest_asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numba==0.61.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai==1.60.1
opencv-python-headless==4.11.0.86
orjson==3.10.15
outlines==0.1.11
outlines_core==0.1.26
packaging==24.2
pandas==2.2.3
parso==0.8.4
partial-json-parser==0.2.1.1.post5
peft==0.12.0
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.1.0
pip==25.0.1
platformdirs==4.3.6
pluggy==1.5.0
pooch==1.8.2
prometheus_client==0.21.1
prometheus-fastapi-instrumentator==7.0.2
prompt_toolkit==3.0.50
propcache==0.3.0
protobuf==3.20.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==19.0.1
pybind11==2.13.6
pycountry==24.6.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.0
ray==2.42.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rouge-chinese==1.0.3
rpds-py==0.22.3
ruff==0.9.9
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
semantic-version==2.10.0
sentencepiece==0.2.0
setuptools==75.1.0
shellingham==1.5.4
shtab==1.7.1
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
sse-starlette==2.2.1
stack_data==0.6.3
starlette==0.46.0
sympy==1.13.1
tabulate==0.9.0
termcolor==2.5.0
threadpoolctl==3.5.0
tiktoken==0.9.0
together==1.3.14
tokenizers==0.21.0
tomlkit==0.13.2
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
triton==3.1.0
trl==0.9.6
typeguard==4.4.2
typer==0.15.2
typing_extensions==4.12.2
tyro==0.8.14
tzdata==2025.1
unsloth==2024.11.6
unsloth_zoo==2025.3.1
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
vllm==0.6.5
watchfiles==1.0.4
wcwidth==0.2.13
websockets==14.2
wheel==0.45.1
xformers==0.0.28.post3
xgrammar==0.1.11
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0